motor test

+++ robot.py
@@ -164,7 +164,7 @@ class Robot:
             integral += error
             derivative = error - last_error
 
-            speed = Kp * error + Ki * integral + Kd * derivative
+            speed = int(Kp * error + Ki * integral + Kd * derivative)
             if speed > 1050:
                 speed = 1050
             elif speed < -1050:
